requirements.txt
crewai
langchain-openai
python-dotenv
fastapi
uvicorn[standard]


